// projects/ui-zanobi/src/lib/znb-base/znb-base.types.ts
export type keyBase =
  | 'Custom'
  | 'Vertical'
  | 'Horizontal'
  | 'Flat'
  | 'Fill'
  | 'Outline'
  | 'Round'
  | 'Pill'
  | 'Raised'
  | 'Primary'
  | 'Secondary'
  | 'Success'
  | 'Warning'
  | 'Danger'
  | 'Info'
  | 'Color'
  | 'Variant'
  | 'Segmented';

export type Colors =
  | 'Primary'
  | 'Secondary'
  | 'Success'
  | 'Danger'
  | 'Warning'
  | 'Info';